backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.28
langchain-anthropic>=0.3.0
anthropic>=0.49.0
openai>=1.50.0
assemblyai>=0.30.0
celery[redis]>=5.4.0
redis>=5.1.0
psycopg2-binary>=2.9.10
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
yt-dlp>=2024.11.18
pydantic>=2.9.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
httpx>=0.27.0
